requirements.txt
openpyxl>=3.1.0
Pillow>=10.0.0
tkinterdnd2>=0.4.0
pyinstaller>=6.0.0
